small refactor
