+++ src/pages/OffersPage/components/PendingOffersTable/hooks/index.ts
@@ -1,0 +1,2 @@
+export * from './usePendingOfferTable'
+export * from './useUserOffers'
