+++ src/pages/OffersPage/components/PendingOffersTable/hooks/usePendingOfferTable.ts
@@ -1,39 +1,17 @@
 import { useMemo, useState } from 'react'
 
-import { useWallet } from '@solana/wallet-adapter-react'
-import { useQuery } from '@tanstack/react-query'
 import { first, groupBy, map } from 'lodash'
 
 import { SearchSelectProps } from '@banx/components/SearchSelect'
 import { SortOption } from '@banx/components/SortDropdown'
 import { createSolValueJSX } from '@banx/components/TableComponents'
 
-import { fetchUserOffers } from '@banx/api/core'
 import { DEFAULT_SORT_OPTION } from '@banx/pages/LoansPage/constants'
 import { calculateLoanValue } from '@banx/utils'
 
-export const useUserOffers = () => {
-  const { publicKey } = useWallet()
-  const publicKeyString = publicKey?.toBase58() || ''
+import { useUserOffers } from './useUserOffers'
 
-  const { data, isLoading } = useQuery(
-    ['userOffers', publicKeyString],
-    () => fetchUserOffers({ walletPublicKey: publicKeyString }),
-    {
-      enabled: !!publicKeyString,
-      staleTime: 5 * 1000,
-      refetchOnWindowFocus: false,
-      refetchInterval: 15 * 1000,
-    },
-  )
-
-  return {
-    offers: data ?? [],
-    loading: isLoading,
-  }
-}
-
-export interface SearchSelectOption {
+interface SearchSelectOption {
   collectionName: string
   collectionImage: string
   bestOffer: number
